+++ salesapp/src/com/isecinc/pens/bean/ReceiptBy.java
@@ -32,8 +32,8 @@ public class ReceiptBy extends I_PO {
 	 * @param rst
 	 */
 	public ReceiptBy(ResultSet rst) throws Exception {
-		setId(rst.getInt("RECEIPT_BY_ID"));
-		setReceiptId(rst.getInt("RECEIPT_ID"));
+		setId(rst.getLong("RECEIPT_BY_ID"));
+		setReceiptId(rst.getLong("RECEIPT_ID"));
 		setPaymentMethod(rst.getString("PAYMENT_METHOD"));
 		setReceiptAmount(rst.getDouble("RECEIPT_AMOUNT"));
 		setCreditCardType(ConvertNullUtil.convertToString(rst.getString("CREDIT_CARD_TYPE")).trim());
@@ -91,10 +91,10 @@ public class ReceiptBy extends I_PO {
 	}
 
 	/** RECEIPT_BY_ID */
-	private int id;
+	private long id;
 
 	/** RECEIPT_ID */
-	private int receiptId;
+	private long receiptId;
 
 	/** PAYMENT_METHOD */
 	private String paymentMethod;
@@ -161,19 +161,19 @@ public class ReceiptBy extends I_PO {
 		this.receiveCashDate = receiveCashDate;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 
-	public int getReceiptId() {
+	public long getReceiptId() {
 		return receiptId;
 	}
 
-	public void setReceiptId(int receiptId) {
+	public void setReceiptId(long receiptId) {
 		this.receiptId = receiptId;
 	}
